refactor: make submodule file names private
