+++ moll/__init__.py
@@ -1,13 +1,9 @@
-import sys
+import sys as _sys
 
-import moll.picking
-import moll.typing
-import moll.utils
+from . import metrics, picking, typing, utils  # noqa: F401
 
-__all__ = []
-
-DEBUG = sys.gettrace() is not None
-TEST = "pytest" in sys.modules
+DEBUG = _sys.gettrace() is not None
+TEST = "pytest" in _sys.modules
 JIT = not DEBUG
 
 # JIT = False
