+++ moll/metrics/__init__.py
@@ -1,1 +1,1 @@
-from .metrics import *  # noqa: F403 (import *)
+from ._metrics import *  # noqa: F403 (import *)
