+++ moll/picking/__init__.py
@@ -1,1 +1,1 @@
-from .online_picker import *  # noqa: F403 (import *)
+from ._online_picker import *  # noqa: F403 (import *)
